+++ src/App.jsx
@@ -1,0 +1,13 @@
+import Navbar from "./components/Navbar";
+import Footer from "./components/Footer";
+
+function App() {
+  return (
+    <>
+      <Navbar />
+      <Footer />
+    </>
+  );
+}
+
+export default App;
