+++ src/components/Footer.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Footer() {
+  return <div>Footer</div>;
+}
