+++ src/components/Navbar.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Navbar() {
+  return <div>Navbar</div>;
+}
