+++ src/pages/AnnouncementPool/index.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function AnnoucementPool() {
+  return <div>AnnoucementPool</div>;
+}
