+++ src/pages/CreateAnnouncement/index.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function CreateAnnouncement() {
+  return <div>CreateAnnouncement</div>;
+}
