+++ src/pages/Home/index.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Home() {
+  return <div>Home</div>;
+}
