+++ src/pages/MyAnnouncements/index.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function MyAnnouncements() {
+  return <div>MyAnnouncements</div>;
+}
